Resolve network deadlock
Prior to this change nodes were able to set their "ballot.counter"
without also setting their "ballot.value". Nodes are also forbidden from
sending their prepare message until they have set "ballot.value".

This posed a problem, since a node entering the prepare phase with no
composite value from nomination and no ballot messages from other nodes
would set their "ballot.counter" to 1 but not their "ballot.value" and
therefore would not be able to send their ballot message to the network.

At this point, the only way this node can change their "ballot.counter",
and subsequently send their prepare message to the network is to see a
blocking threshold of nodes with "ballot.counter" greater than theirs or
a quorum of other nodes with "ballot.counter" greater than or equal to
theirs.  Notably even if this node were to confirm nominated a value, it
would not be able to update its prepare message and send it to the
network.

If a sufficient number of other nodes are also not able to send their
prepare message, then it can happen that no nodes in the network are now
able to see a blocking or quorum threshold of messages from other nodes
and the network will get stuck with no nodes able to progress.

This change modifies the conditions for setting "ballot.counter" so that
a node may not set its "ballot.counter" unless it can also set its
"ballot.value". This ensures that as soon as a value is confirmed
nominated, a node will be able to start sending its prepare message.