Prevent messages conveying conflicting statements
It can happen that a node that is in the PREPARE phase with b=(1,a)
moves to the COMMIT phase by accepting comitted c=h=(2,b) and
subsequently accepts comitted (1,b) so c=(1,b) h=(2,b)

If the value b orders before a then during this sequence of operations
the node will have conveyed a vote to abort (1,b) and a vote to commit
(1,b).

I think the solution to this is to modify the statements conveyed by the
SCPCommit message so that only ballots greater than or equal to h are
voted for.